things can now draw if all the settings are just right

+++ opengl-gui/include/Graphics/WidgetDetails.h
@@ -1,7 +1,7 @@
 #pragma once
 namespace OpenGLGUI {
 	struct WidgetDetails {
-		float x, y, width, height;
+		int x, y, width, height;
 	};
 }
